feat: fast migrating to typescript
